requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pillow>=8.3.0
opencv-python>=4.5.0
flask>=2.0.0
tqdm>=4.62.0
albumentations>=1.1.0
seaborn>=0.11.0
streamlit
opencv-python